add controler and router for reports
