+++ controlers/index.ts
@@ -1,2 +1,3 @@
 export * from './user';
 export * from './reminder';
+export * from './report';
